the java websocket can connect to postman
