+++ server/src/main/java/com/codeclan/server/components/MyWebSocketHandler.java
@@ -1,0 +1,32 @@
+package com.codeclan.server.components;
+
+import org.springframework.web.socket.*;
+
+public class MyWebSocketHandler implements WebSocketHandler {
+
+    @Override
+    public void afterConnectionEstablished(WebSocketSession session) throws Exception {
+        System.out.println("WebSocket connection established.");
+    }
+
+    @Override
+    public void handleMessage(WebSocketSession session, WebSocketMessage<?> message) throws Exception {
+        System.out.println("Received message: " + message.getPayload());
+        session.sendMessage(new TextMessage("Hello, client!"));
+    }
+
+    @Override
+    public void handleTransportError(WebSocketSession session, Throwable exception) throws Exception {
+        System.out.println("Transport error: " + exception.getMessage());
+    }
+
+    @Override
+    public void afterConnectionClosed(WebSocketSession session, CloseStatus closeStatus) throws Exception {
+        System.out.println("WebSocket connection closed.");
+    }
+
+    @Override
+    public boolean supportsPartialMessages() {
+        return false;
+    }
+}
