--- server/src/main/java/com/codeclan/server/models/MessageResponse.java
@@ -1,28 +1,0 @@
-package com.codeclan.server.models;
-
-public class MessageResponse {
-    private String from;
-    private String text;
-
-    public MessageResponse(String from, String text) {
-        this.from = from;
-        this.text = text;
-    }
-
-    public String getFrom() {
-        return from;
-    }
-
-    public void setFrom(String from) {
-        this.from = from;
-    }
-
-    public String getText() {
-        return text;
-    }
-
-    public void setText(String text) {
-        this.text = text;
-    }
-}
-
