+++ server/src/main/resources/application.properties
@@ -7,7 +7,7 @@ spring.jpa.hibernate.ddl-auto=update
 spring.jpa.database-platform=org.hibernate.dialect.PostgreSQL9Dialect
 spring.jpa.properties.hibernate.enable_lazy_load_no_trans=true
 #server.ssl.enabled=true
-server.port=8080
+#server.port=8080
 #server.ssl.protocol=TLS
-server.address=192.168.0.5
+#server.address=192.168.0.5
